Dates are now cleaned after record
immediately after extracting the dublin core for a single record, the 
date format is cleaned up before moving on to the next record in the 
folder. If opened in Excel, the standard excel mm/dd/yyyy display format 
is superimposed on complete dates, this does not alter the underlying 
csv, which is correct.
Have also done a little minor refactoring.