tasks to automate  pid identification and record export now working
metadata tasks now in place to 1) run a sparql query which will identify 
theses, edit the result of this query into the correct syntax for input 
to a bulk export script and upload it to the fedora client on a remote 
fedora host 2) a rake task to take the task created at 1) as its input, 
use this to run a bulk export of records from fedora and finally 
transfer them to a suitable location on the box handling the metadata 
extraction  

TODO the rake tasks need tidying up (the separate transfer tasks was for 
development only) and the new tasks integrating into the overall 
metadata creation process. It is important that it should still be 
possible to run the metadata extraction without neccesarily performing a 
complete new record extract.