rake record identification and export taks now chained together
The do_all_extraction_tasks task now runs the get_theses_pids and 
export_records tasks sequentially. there is a strange syntax error which 
does not seem to affect the functionality. TODO I have made the name of 
the list of pids configurable in the export task but not in the pid 
identification - this needs to be made consistent for both